requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]
certifi
scikit-learn
PyYAML
dill
# -e .